+++ lib/activerecord-import/adapters/postgresql_adapter.rb
@@ -5,9 +5,10 @@ module ActiveRecord::Import::PostgreSQLAdapter
   MIN_VERSION_FOR_UPSERT = 90_500
 
   def insert_many( sql, values, options = {}, *args ) # :nodoc:
-    primary_key = options[:primary_key]
     number_of_inserts = 1
+    returned_values = []
     ids = []
+    results = []
 
     base_sql, post_sql = if sql.is_a?( String )
       [sql, '']
@@ -17,11 +18,12 @@ module ActiveRecord::Import::PostgreSQLAdapter
 
     sql2insert = base_sql + values.join( ',' ) + post_sql
 
-    if primary_key.blank? || options[:no_returning]
+    columns = returning_columns(options)
+    if columns.blank? || options[:no_returning]
       insert( sql2insert, *args )
     else
-      ids = if primary_key.is_a?( Array )
-        # Select composite primary keys
+      returned_values = if columns.size > 1
+        # Select composite columns
         select_rows( sql2insert, *args )
       else
         select_values( sql2insert, *args )
@@ -29,7 +31,34 @@ module ActiveRecord::Import::PostgreSQLAdapter
       query_cache.clear if query_cache_enabled
     end
 
-    [number_of_inserts, ids]
+    if options[:returning].blank?
+      ids = returned_values
+    elsif options[:primary_key].blank?
+      results = returned_values
+    else
+      # split primary key and returning columns
+      ids, results = split_ids_and_results(returned_values, columns, options)
+    end
+
+    ActiveRecord::Import::Result.new([], number_of_inserts, ids, results)
+  end
+
+  def split_ids_and_results(values, columns, options)
+    ids = []
+    results = []
+    id_indexes = Array(options[:primary_key]).map { |key| columns.index(key) }
+    returning_indexes = Array(options[:returning]).map { |key| columns.index(key) }
+
+    values.each do |value|
+      value_array = Array(value)
+      ids << id_indexes.map { |i| value_array[i] }
+      results << returning_indexes.map { |i| value_array[i] }
+    end
+
+    ids.map!(&:first) if id_indexes.size == 1
+    results.map!(&:first) if returning_indexes.size == 1
+
+    [ids, results]
   end
 
   def next_value_for_sequence(sequence_name)
@@ -50,12 +79,19 @@ module ActiveRecord::Import::PostgreSQLAdapter
 
     sql += super(table_name, options)
 
-    unless options[:no_returning] || options[:primary_key].blank?
-      primary_key = Array(options[:primary_key])
-      sql << " RETURNING #{primary_key.join(', ')}"
+    columns = returning_columns(options)
+    unless columns.blank? || options[:no_returning]
+      sql << " RETURNING \"#{columns.join('", "')}\""
     end
 
     sql
+  end
+
+  def returning_columns(options)
+    columns = []
+    columns += Array(options[:primary_key]) if options[:primary_key].present?
+    columns |= Array(options[:returning]) if options[:returning].present?
+    columns
   end
 
   # Add a column to be updated on duplicate key update
@@ -87,10 +123,11 @@ module ActiveRecord::Import::PostgreSQLAdapter
     arg = { columns: arg } if arg.is_a?( Array ) || arg.is_a?( String )
     return unless arg.is_a?( Hash )
 
-    sql = " ON CONFLICT "
+    sql = ' ON CONFLICT '
     conflict_target = sql_for_conflict_target( arg )
 
     columns = arg.fetch( :columns, [] )
+    condition = arg[:condition]
     if columns.respond_to?( :empty? ) && columns.empty?
       return sql << "#{conflict_target}DO NOTHING"
     end
@@ -110,6 +147,9 @@ module ActiveRecord::Import::PostgreSQLAdapter
     else
       raise ArgumentError, 'Expected :columns to be an Array or Hash'
     end
+
+    sql << " WHERE #{condition}" if condition.present?
+
     sql
   end
 
@@ -137,7 +177,7 @@ module ActiveRecord::Import::PostgreSQLAdapter
     if constraint_name.present?
       "ON CONSTRAINT #{constraint_name} "
     elsif conflict_target.present?
-      '(' << Array( conflict_target ).reject( &:empty? ).join( ', ' ) << ') '.tap do |sql|
+      '(' << Array( conflict_target ).reject( &:blank? ).join( ', ' ) << ') '.tap do |sql|
         sql << "WHERE #{index_predicate} " if index_predicate
       end
     end
@@ -157,7 +197,7 @@ module ActiveRecord::Import::PostgreSQLAdapter
     current_version >= MIN_VERSION_FOR_UPSERT
   end
 
-  def support_setting_primary_key_of_imported_objects?
+  def supports_setting_primary_key_of_imported_objects?
     true
   end
 end