+++ test/import_test.rb
@@ -48,6 +48,14 @@ describe "#import" do
           Widget.import Build(3, :widgets)
         end
       end
+
+      context "with uppercase letters" do
+        it "should import models successfully" do
+          assert_difference "Car.count", +3 do
+            Car.import Build(3, :cars)
+          end
+        end
+      end
     end
 
     context "that have no primary key" do
@@ -78,23 +86,73 @@ describe "#import" do
         assert_nil t.author_email_address
       end
     end
-  end
 
-  describe "with composite primary keys" do
-    it "should import models successfully" do
-      tags = [Tag.new(tag_id: 1, publisher_id: 1, tag: 'Mystery')]
+    context "with extra keys" do
+      let(:values) do
+        [
+          { title: "LDAP", author_name: "Jerry Carter" },
+          { title: "Rails Recipes", author_name: "Chad Fowler", author_email_address: "cfowler@test.com" } # author_email_address is unknown
+        ]
+      end
 
-      assert_difference "Tag.count", +1 do
-        Tag.import tags
+      it "should fail when column names are not specified" do
+        err = assert_raises ArgumentError do
+          Topic.import values, validate: false
+        end
+
+        assert err.message.include? 'Extra keys: [:author_email_address]'
+      end
+
+      it "should succeed when column names are specified" do
+        assert_difference "Topic.count", +2 do
+          Topic.import columns, values, validate: false
+        end
       end
     end
 
-    it "should import array of values successfully" do
-      columns = [:tag_id, :publisher_id, :tag]
-      values = [[1, 1, 'Mystery'], [2, 1, 'Science']]
+    context "with missing keys" do
+      let(:values) do
+        [
+          { title: "LDAP", author_name: "Jerry Carter" },
+          { title: "Rails Recipes" } # author_name is missing
+        ]
+      end
 
-      assert_difference "Tag.count", +2 do
-        Tag.import columns, values, validate: false
+      it "should fail when column names are not specified" do
+        err = assert_raises ArgumentError do
+          Topic.import values, validate: false
+        end
+
+        assert err.message.include? 'Missing keys: [:author_name]'
+      end
+
+      it "should fail on missing hash key from specified column names" do
+        err = assert_raises ArgumentError do
+          Topic.import %i(author_name), values, validate: false
+        end
+
+        assert err.message.include? 'Missing keys: [:author_name]'
+      end
+    end
+  end
+
+  unless ENV["SKIP_COMPOSITE_PK"]
+    describe "with composite primary keys" do
+      it "should import models successfully" do
+        tags = [Tag.new(tag_id: 1, publisher_id: 1, tag: 'Mystery')]
+
+        assert_difference "Tag.count", +1 do
+          Tag.import tags
+        end
+      end
+
+      it "should import array of values successfully" do
+        columns = [:tag_id, :publisher_id, :tag]
+        values = [[1, 1, 'Mystery'], [2, 1, 'Science']]
+
+        assert_difference "Tag.count", +2 do
+          Tag.import columns, values, validate: false
+        end
       end
     end
   end
@@ -111,12 +169,12 @@ describe "#import" do
   end
 
   context "with :validation option" do
-    let(:columns) { %w(title author_name) }
-    let(:valid_values) { [["LDAP", "Jerry Carter"], ["Rails Recipes", "Chad Fowler"]] }
-    let(:valid_values_with_context) { [[1111, "Jerry Carter"], [2222, "Chad Fowler"]] }
-    let(:invalid_values) { [["The RSpec Book", ""], ["Agile+UX", ""]] }
-    let(:valid_models) { valid_values.map { |title, author_name| Topic.new(title: title, author_name: author_name) } }
-    let(:invalid_models) { invalid_values.map { |title, author_name| Topic.new(title: title, author_name: author_name) } }
+    let(:columns) { %w(title author_name content) }
+    let(:valid_values) { [["LDAP", "Jerry Carter", "Putting Directories to Work."], ["Rails Recipes", "Chad Fowler", "A trusted collection of solutions."]] }
+    let(:valid_values_with_context) { [[1111, "Jerry Carter", "1111"], [2222, "Chad Fowler", "2222"]] }
+    let(:invalid_values) { [["The RSpec Book", "David Chelimsky", "..."], ["Agile+UX", "", "All about Agile in UX."]] }
+    let(:valid_models) { valid_values.map { |title, author_name, content| Topic.new(title: title, author_name: author_name, content: content) } }
+    let(:invalid_models) { invalid_values.map { |title, author_name, content| Topic.new(title: title, author_name: author_name, content: content) } }
 
     context "with validation checks turned off" do
       it "should import valid data" do
@@ -151,6 +209,22 @@ describe "#import" do
         end
       end
 
+      it "should ignore uniqueness validators" do
+        Topic.import columns, valid_values, validate: true
+        assert_difference "Topic.count", +2 do
+          Topic.import columns, valid_values, validate: true
+        end
+      end
+
+      it "should not alter the callback chain of the model" do
+        attributes = columns.zip(valid_values.first).to_h
+        topic = Topic.new attributes
+        Topic.import [topic], validate: true
+        duplicate_topic = Topic.new attributes
+        Topic.import [duplicate_topic], validate: true
+        assert duplicate_topic.invalid?
+      end
+
       it "should not import invalid data" do
         assert_no_difference "Topic.count" do
           Topic.import columns, invalid_values, validate: true
@@ -174,7 +248,7 @@ describe "#import" do
       end
 
       it "should set ids in valid models if adapter supports setting primary key of imported objects" do
-        if ActiveRecord::Base.support_setting_primary_key_of_imported_objects?
+        if ActiveRecord::Base.supports_setting_primary_key_of_imported_objects?
           Topic.import (invalid_models + valid_models), validate: true
           assert_nil invalid_models[0].id
           assert_nil invalid_models[1].id
@@ -184,7 +258,7 @@ describe "#import" do
       end
 
       it "should set ActiveRecord timestamps in valid models if adapter supports setting primary key of imported objects" do
-        if ActiveRecord::Base.support_setting_primary_key_of_imported_objects?
+        if ActiveRecord::Base.supports_setting_primary_key_of_imported_objects?
           Timecop.freeze(Time.at(0)) do
             Topic.import (invalid_models + valid_models), validate: true
           end
@@ -206,6 +280,18 @@ describe "#import" do
           Topic.import columns, valid_values + invalid_values, validate: true
         end
         assert_equal 0, Topic.where(title: invalid_values.map(&:first)).count
+      end
+
+      it "should run callbacks" do
+        assert_no_difference "Topic.count" do
+          Topic.import columns, [["invalid", "Jerry Carter"]], validate: true
+        end
+      end
+
+      it "should call validation methods" do
+        assert_no_difference "Topic.count" do
+          Topic.import columns, [["validate_failed", "Jerry Carter"]], validate: true
+        end
       end
     end
   end
@@ -282,7 +368,7 @@ describe "#import" do
 
       it "doesn't reload any data (doesn't work)" do
         Topic.import new_topics, synchronize: new_topics
-        if Topic.support_setting_primary_key_of_imported_objects?
+        if Topic.supports_setting_primary_key_of_imported_objects?
           assert new_topics.all?(&:persisted?), "Records should have been reloaded"
         else
           assert new_topics.all?(&:new_record?), "No record should have been reloaded"
@@ -463,7 +549,7 @@ describe "#import" do
 
   context "with has_and_belongs_to_many" do
     it "should import has_and_belongs_to_many ids also" do
-      if ActiveRecord::Base.support_setting_primary_key_of_imported_objects?
+      if ActiveRecord::Base.supports_setting_primary_key_of_imported_objects?
         Genre.import [:title], [['Horror'], ['Romance'], ['Fiction']]
         Book.import  [Book.new(title: "Fell", author_name: "Curry", publisher: "Bayer", created_at: 2.years.ago.utc, created_on: 2.years.ago.utc, genres: [Genre.first, Genre.last])]
 
@@ -514,7 +600,24 @@ describe "#import" do
 
           assert_equal [val1, val2], scope.map(&column).sort
         end
+
+        it "works importing array of hashes" do
+          scope.import [{ column => val1 }, { column => val2 }]
+
+          assert_equal [val1, val2], scope.map(&column).sort
+        end
       end
+    end
+  end
+
+  context "importing model with polymorphic belongs_to" do
+    it "works without error" do
+      book     = FactoryGirl.create :book
+      discount = Discount.new(discountable: book)
+
+      Discount.import([discount])
+
+      assert_equal 1, Discount.count
     end
   end
 
@@ -592,6 +695,30 @@ describe "#import" do
     end
   end
 
+  context 'importing arrays of values with boolean fields' do
+    let(:columns) { [:author_name, :title, :for_sale] }
+
+    it 'should be able to coerce integers as boolean fields' do
+      Book.delete_all if Book.count > 0
+      values = [['Author #1', 'Book #1', 0], ['Author #2', 'Book #2', 1]]
+      assert_difference "Book.count", +2 do
+        Book.import columns, values
+      end
+      assert_equal false, Book.first.for_sale
+      assert_equal true, Book.last.for_sale
+    end
+
+    it 'should be able to coerce strings as boolean fields' do
+      Book.delete_all if Book.count > 0
+      values = [['Author #1', 'Book #1', 'false'], ['Author #2', 'Book #2', 'true']]
+      assert_difference "Book.count", +2 do
+        Book.import columns, values
+      end
+      assert_equal false, Book.first.for_sale
+      assert_equal true, Book.last.for_sale
+    end
+  end
+
   describe "importing when model has default_scope" do
     it "doesn't import the default scope values" do
       assert_difference "Widget.unscoped.count", +2 do
@@ -639,6 +766,16 @@ describe "#import" do
         Widget.import [:w_id, :json_data], [[9, data]]
       end
       assert_equal(data.as_json, Widget.find_by_w_id(9).json_data)
+    end
+
+    it "imports serialized values from saved records" do
+      Widget.import [:w_id, :json_data], [[1, data]]
+      assert_equal data.as_json, Widget.last.json_data
+
+      w = Widget.last
+      w.w_id = 2
+      Widget.import([w])
+      assert_equal data.as_json, Widget.last.json_data
     end
 
     context "with a store" do
@@ -696,6 +833,20 @@ describe "#import" do
           assert_difference "Topic.count", +2 do
             Topic.import! columns, valid_values
           end
+        end
+      end
+    end
+
+    context "with objects that respond to .to_sql as values" do
+      let(:columns) { %w(title author_name) }
+      let(:valid_values) { [["LDAP", Book.select("'Jerry Carter'").limit(1)], ["Rails Recipes", Book.select("'Chad Fowler'").limit(1)]] }
+
+      it "should import data" do
+        assert_difference "Topic.count", +2 do
+          Topic.import! columns, valid_values
+          topics = Topic.all
+          assert_equal "Jerry Carter", topics.first.author_name
+          assert_equal "Chad Fowler", topics.last.author_name
         end
       end
     end